add support for rspec 3
