+++ lib/rainbow.rb
@@ -1,35 +1,9 @@
-require 'rspec/core/formatters/progress_formatter'
-
-class Rainbow < ::RSpec::Core::Formatters::ProgressFormatter
-  PI_3 = Math::PI / 3
-
-  def initialize(options)
-    # colors calculation stolen from Minitest's Pride Plugin
-    # https://github.com/seattlerb/minitest
-    @colors = (0...(6 * 7)).map { |n|
-      n *= 1.0 / 6
-      r  = (3 * Math.sin(n           ) + 3).to_i
-      g  = (3 * Math.sin(n + 2 * PI_3) + 3).to_i
-      b  = (3 * Math.sin(n + 4 * PI_3) + 3).to_i
-
-      36 * r + 6 * g + b + 16
-    }
-    @color_index = 0
-    super(options)
-  end
-
-  def rainbow
-    @color_index == (@colors.size - 1) ? @color_index = 0 : @color_index += 1
-    @colors[@color_index]
-  end
-
-  def color(text, color_code)
-    color_code = rainbow if text == '.'
-    super
-  end
-
-  def color_code_for(code_or_symbol)
-    return "38;5;#{code_or_symbol}" if code_or_symbol.is_a?(Integer)
-    super
-  end
+if Gem::Version.new(RSpec::Core::Version::STRING).release >= Gem::Version.new('3.1.0')
+  require 'rainbow_formatter/rspec3'
+  formatter = RSpec3
+else
+  require 'rainbow_formatter/rspec2'
+  formatter = RSpec2
 end
+
+Rainbow = formatter
